Features: added clickable tile grid, clicking turns grid red. Grid no longer resizeable so won't work on phones

+++ src/shared/components.rs
@@ -1,7 +1,7 @@
 use bevy::ecs::component::Component;
 
 #[derive(Component)]
-pub struct Health(pub i32);
+pub struct Health(pub f32);
 
 #[derive(Component)]
-pub struct Damage(pub i32);
+pub struct Damage(pub f32);
